requirements.txt
django>=2.0,<2.1
django-model-utils>=2.0
# Additional requirements go here
git+https://github.com/darklow/django-suit.git@v2 # django-suit>=2.0a1
django-crispy-forms>=1.5.0
arrow>=0.12
